% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateChipSeqPeaks.R
\name{annotateChipSeqPeaks}
\alias{annotateChipSeqPeaks}
\title{A function to match each peak to associated genes in terms of RefID.}
\usage{
annotateChipSeqPeaks(chip.seq, transcriptDB = NULL,
  distanceRange = c(-1e+06, 1e+06))
}
\arguments{
\item{chip.seq}{A matrix listing all CHIP-Seq peaks in which the chromosome
ID for all peaks are listed in the first column, peaks' start
positions are listed in the second column, peaks' end
position are listed in the third column, peaks' ID are listed
in the forth column and the fifth column lists peaks' score.}

\item{transcriptDB}{This is Transcript definition table created by
makeTranscriptDbFromUCSC or makeTxDbFromUCSC functions.}

\item{distanceRange}{The up- and downstream window aroung gene TSS to be used
in associating peaks to a gene. The default is c(-1e+06,1e+06).}
}
\value{
This function returns a matrix in which the first column contains
the RefSeqID, the second column contains Entrez ID , the third
column contains chromosome ID, the fourth column contains the
absolute distance from each peak to transcription start site and
the fifth column contains chip-seq score.
}
\description{
This function allows you to annotate chip-sequencing reads to a reference.
It is intended to search all peaks within up and down
distances for each gene. Users could specify their own platforms
or the function will download platform from UCSC as default.
}
\details{
This function is intended to search all peaks within up and down distances for each gene.
}
\examples{
## not run
data(erAlpha)
refTable <- GenomicFeatures::makeTxDbFromUCSC(genome="hg19",tablename="refGene")
ChipSeq <- annotateChipSeqPeaks(chip.seq=erAlpha[[3]], transcriptDB=refTable, distanceRange=c(-1e+06,1e+06))
## not run
}
\author{
Mehdi Fazel-Najafabadi, Mario Medvedovic
}
